+++ src/controllers/authController.ts
@@ -3,7 +3,7 @@ import { SigninData, SignupData } from '../contracts/auth';
 import API, { AuthAPI } from '../api/AuthAPI';
 import router from '../utils/Router';
 import { Routes } from '..';
-import storage from '../utils/Store';
+import { storage } from '../utils/Store';
 
 export class AuthController {
   private readonly api: AuthAPI;
