+++ src/utils/StoreTypes.ts
@@ -1,0 +1,14 @@
+import { UserInfo } from '../api/ChatsAPI';
+import { User } from '../contracts/auth';
+import { Message } from '../contracts/message';
+
+export enum StoreEvents {
+  Updated = 'updated',
+}
+
+export interface State {
+  user: User;
+  usersData: UserInfo[];
+  messages: Record<number, Message[]>;
+  selectedChat?: number;
+}
