+++ src/utils/withStore.ts
@@ -1,9 +1,14 @@
 /* eslint-disable import/no-named-as-default */
-import storage, { State, StoreEvents } from './Store';
+import { storage } from './Store';
 import BlockConstructable from './Block';
+import { State, StoreEvents } from './StoreTypes';
 
-export function withStore<SP extends Partial<Record<string, any>>>(mapStateToProps: (state: State) => SP) {
-  return function wrap<P>(Component: typeof BlockConstructable<Record<string, any>>) {
+export function withStore<SP extends Partial<Record<string, any>>>(
+  mapStateToProps: (state: State) => SP,
+) {
+  return function wrap<P>(
+    Component: typeof BlockConstructable<Record<string, any>>,
+  ) {
     return class WithStore extends Component {
       constructor(props: Omit<P, keyof SP>) {
         let previousState = mapStateToProps(storage.getState());
